fix(DPI): expanded full width of 'source' and 'sink' (#25)

+++ main/DPI/tilelink_dpi.hpp
@@ -57,7 +57,7 @@ extern "C" void TileLinkPullChannelA(
     uint8_t*            opcode,
     uint8_t*            param,
     uint8_t*            size,
-    uint8_t*            source,
+    uint64_t*           source,
     uint64_t*           address,
     uint8_t*            user_needHint,
     uint64_t*           user_vaddr,
@@ -80,7 +80,7 @@ extern "C" void TileLinkPushChannelB(
     const uint8_t       opcode,
     const uint8_t       param,
     const uint8_t       size,
-    const uint8_t       source,
+    const uint64_t      source,
     const uint64_t      address,
     const uint32_t      mask,
     const uint64_t      data0,
@@ -108,7 +108,7 @@ extern "C" void TileLinkPullChannelC(
     uint8_t*            opcode,
     uint8_t*            param,
     uint8_t*            size,
-    uint8_t*            source,
+    uint64_t*           source,
     uint64_t*           address,
     uint8_t*            user_needHint,
     uint64_t*           user_vaddr,
@@ -130,8 +130,8 @@ extern "C" void TileLinkPushChannelD(
     const uint8_t       opcode,
     const uint8_t       param,
     const uint8_t       size,
-    const uint8_t       source,
-    const uint8_t       sink,
+    const uint64_t      source,
+    const uint64_t      sink,
     const uint8_t       denied,
     const uint64_t      data0,
     const uint64_t      data1,
@@ -155,7 +155,7 @@ extern "C" void TileLinkPushChannelE(
 extern "C" void TileLinkPullChannelE(
     const int           deviceId,
     uint8_t*            valid,
-    uint8_t*            sink);
+    uint64_t*           sink);
 //
